Multimodal_Agent/requirements.txt
google-generativeai>=0.5.0
pyautogui>=0.9.54
Pillow>=10.0.0
pygetwindow>=0.0.9
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
